seed on backtest
