+++ lib/bt/fetch_data.js
@@ -8,13 +8,13 @@ const { DEFAULT_ITEMS_LIMIT } = require('./constants')
  * @return {function(*, *): *}
  */
 const fetchCandles = (rest, semaphore, { symbol, timeframe }) => {
-  return (offset, end) => {
+  return (offset, end, limit = DEFAULT_ITEMS_LIMIT) => {
     return semaphore.add(
       rest.candles.bind(rest, {
         timeframe,
         symbol,
         query: {
-          limit: DEFAULT_ITEMS_LIMIT,
+          limit,
           start: offset,
           end,
           sort: 1
