+++ lib/cmds/exec_strategy.js
@@ -2,7 +2,7 @@
 
 const debug = require('debug')('bfx:hf:data-server:cmds:exec-bt')
 
-const { rangeString, candleWidth } = require('bfx-hf-util')
+const { rangeString } = require('bfx-hf-util')
 const { execOffline } = require('bfx-hf-backtest')
 const HFS = require('bfx-hf-strategy')
 const Indicators = require('bfx-hf-indicators')
@@ -14,12 +14,15 @@ const sendError = require('../wss/send_error')
 const send = require('../wss/send')
 const DataPointFeed = require('../bt/data_feed')
 const DataPointStream = require('../bt/data_stream')
-const { fetchCandles, fetchTrades } = require('../bt/fetch_data')
+const {
+  fetchCandles: fetchCandlesFactory,
+  fetchTrades: fetchTradesFactory
+} = require('../bt/fetch_data')
 const ExecutionContext = require('../bt/context')
 const RequestSemaphore = require('../bt/request_semaphore')
+const seedCandlesFactory = require('../bt/seed_candles')
 const generateResults = require('bfx-hf-strategy/lib/util/generate_strategy_results')
 const parseStrategy = require('bfx-hf-strategy/lib/util/parse_strategy')
-const { onSeedCandle } = require('bfx-hf-strategy')
 
 /**
  * @param {DataServer} ds
@@ -70,6 +73,7 @@ module.exports = async (ds, ws, msg) => {
   const dataPointFeed = new DataPointFeed()
   const rest = new RESTv2({ transform: true })
   const requestSemaphore = new RequestSemaphore()
+  let fetchCandles, seedCandles
 
   debug(
     'running backtest for %s:%s [%s]',
@@ -77,37 +81,25 @@ module.exports = async (ds, ws, msg) => {
   )
 
   if (includeCandles) {
-    const fetchStrategy = fetchCandles(rest, requestSemaphore, { symbol, timeframe })
-    const stream = new DataPointStream(fetchStrategy)
+    fetchCandles = fetchCandlesFactory(rest, requestSemaphore, { symbol, timeframe })
+    const stream = new DataPointStream(fetchCandles)
     dataPointFeed.addStream(stream)
   }
 
   if (includeTrades) {
-    const fetchStrategy = fetchTrades(rest, requestSemaphore, { symbol })
+    const fetchStrategy = fetchTradesFactory(rest, requestSemaphore, { symbol })
     const stream = new DataPointStream(fetchStrategy)
     dataPointFeed.addStream(stream)
   }
 
   if (candleSeed) {
-    const cWidth = candleWidth(timeframe)
-    const seedStart = start - (candleSeed * cWidth)
-    const candles = await rest.candles({
-      timeframe,
+    seedCandles = seedCandlesFactory({
       symbol,
-      query: {
-        limit: candleSeed,
-        start: seedStart,
-        end: start,
-        sort: 1
-      }
+      timeframe,
+      fetchCandles,
+      start,
+      candleSeed
     })
-
-    for (const candle of candles) {
-      candle.tf = timeframe
-      candle.symbol = symbol
-
-      strategy = onSeedCandle(strategy, candle)
-    }
   }
 
   ds.activeBacktests.set(strategy.gid, context)
@@ -117,6 +109,8 @@ module.exports = async (ds, ws, msg) => {
     end,
     includeTrades,
     includeCandles,
+    candleSeed,
+    seedCandles,
     priceFeed,
     perfManager,
     startPerformanceWatchers,
