+++ package.json
@@ -38,7 +38,7 @@
   "dependencies": {
     "bfx-api-node-rest": "^4.6.0",
     "bfx-api-node-util": "^1.0.9",
-    "bfx-hf-backtest": "git+https://github.com/tarcisiozf/bfx-hf-backtest.git#fix-trades-list",
+    "bfx-hf-backtest": "git+https://github.com/tarcisiozf/bfx-hf-backtest.git#seed-candles",
     "bfx-hf-indicators": "git+https://github.com/bitfinexcom/bfx-hf-indicators.git#v2.1.1",
     "bfx-hf-models": "git+https://github.com/bitfinexcom/bfx-hf-models.git#v3.0.0",
     "bfx-hf-strategy": "git+https://github.com/bitfinexcom/bfx-hf-strategy.git#v1.5.0",
